+++ routes/api.php
@@ -1,21 +1,15 @@
 <?php
 
-$resourcePathPattern = '[0-9a-zA-Z-_@&\#\!=,:;\/\^\$\.\|\{\}\[\]\(\)\*\+\? ]+';
+$versionPattern = 'v[0-9.]+';
 $servicePattern = '[_0-9a-zA-Z-.]+';
+$resourcePathPattern = '[0-9a-zA-Z-_@&\#\!=,:;\/\^\$\.\|\{\}\[\]\(\)\*\+\? ]+';
 
-Route::get('{version}/', 'RestController@index');
-Route::get('{version}/{service}/{resource?}', 'RestController@handleGET')->where(
-    ['service' => $servicePattern, 'resource' => $resourcePathPattern]
+Route::get('', 'RestController@index');
+// Support old versioning in URL, i.e api/v2 and api/v2/service
+Route::get('{version}', 'RestController@index')->where(['version' => $versionPattern]);
+Route::any('{version}/{service}/{resource?}', 'RestController@handleVersionedService')->where(
+    ['version' => $versionPattern, 'service' => $servicePattern, 'resource' => $resourcePathPattern]
 );
-Route::post('{version}/{service}/{resource?}', 'RestController@handlePOST')->where(
-    ['service' => $servicePattern, 'resource' => $resourcePathPattern]
-);
-Route::put('{version}/{service}/{resource?}', 'RestController@handlePUT')->where(
-    ['service' => $servicePattern, 'resource' => $resourcePathPattern]
-);
-Route::patch('{version}/{service}/{resource?}', 'RestController@handlePATCH')->where(
-    ['service' => $servicePattern, 'resource' => $resourcePathPattern]
-);
-Route::delete('{version}/{service}/{resource?}', 'RestController@handleDELETE')->where(
+Route::any('{service}/{resource?}', 'RestController@handleService')->where(
     ['service' => $servicePattern, 'resource' => $resourcePathPattern]
 );